# Extend interacting pairs