--- a/src/core/continuous_space.jl
+++ b/src/core/continuous_space.jl
@@ -315,11 +315,6 @@ function elastic_collision!(a, b, f = nothing)
   return true
 end
 
-# TODO: Add "true_interacting_pairs" where it is guaranteed that each agent
-# is only paired with its true nearest neighbor
-
-# TODO: add "all_interacting_pairs", the original function asked in #220
-
 """
     interacting_pairs(model, r, scheduler = model.scheduler)
 Return an iterator that yields unique pairs of agents `(a1, a2)` that are closest
@@ -331,46 +326,20 @@ some pairwise interaction across all pairs of closest agents once
 (and does not want to trigger the event twice, both with `a1` and with `a2`, which
 is unavoidable when using `agent_step!`).
 
-Notice that by default the pairs are created by scanning each agent according to the given
-`scheduler`. This is important, because this function does not match each agent with
-its absolute nearest neighbor. Imagine three agents A B C where the
-nearest neighbor of A is B but the nearest neighbor of B is C. If you start with A,
-you get the pair (A, B), but if you start with B you get (B, C).
-
-The keyword argument `method = :scheduler` provides two additional pairing scenarios
-- `:true`: where it is guaranteed that each agent is only paired with its true
-nearest neighbor.
-- `:all`: returns every pair of agents within the interaction radius `r` regardless of
-their nearest neighbor status.
+The keyword argument `all = false` can be set `true` to return every pair of agents
+within the interaction radius `r` regardless of their nearest neighbor status.
 """
-function interacting_pairs(model, r, scheduler = model.scheduler; method = :scheduler)
-  @assert method ∈ (:scheduler, :true, :all)
+function interacting_pairs(model::ABM, r::Real; all = false)
   pairs = Tuple{Int, Int}[]
-  if method == :scheduler
-    scheduler_pairs!(pairs, model, r, scheduler)
-  elseif method == :all
-    all_pairs!(pairs, model, r)
-  else
+  if !all
     true_pairs!(pairs, model, r)
+  else
+    all_pairs!(pairs, model, r)
   end
   return PairIterator(pairs, model.agents)
 end
 
-function scheduler_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r, scheduler)
-  #TODO: This can be optimized further I assume
-  for id in scheduler(model)
-    # Skip already checked agents
-    any(isequal(id), p[2] for p in pairs) && continue
-    a1 = model[id]
-    a2 = nearest_neighbor(a1, model, r)
-    # This line ensures each neighbor exists in only one pair:
-    if a2 ≠ nothing && !any(isequal(a2.id), p[2] for p in pairs)
-      push!(pairs, (id, a2.id))
-    end
-  end
-end
-
-function all_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r)
+function all_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r::Real)
   for a in allagents(model)
     for nid in space_neighbors(a, model, r)
       # Sort the pair to overcome any uniqueness issues
@@ -380,18 +349,18 @@ function all_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r)
   end
 end
 
-function true_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r)
+function true_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r::Real)
   distances = Vector{Float64}(undef, 0)
   for a in allagents(model)
-    nid = nearest_neighbor(a, model, r).id
-    nid == nothing && break
+    nn = nearest_neighbor(a, model, r)
+    nn == nothing && break
     # Sort the pair to overcome any uniqueness issues
-    new_pair = isless(a.id, nid) ? (a.id, nid) : (nid, a.id)
+    new_pair = isless(a.id, nn.id) ? (a.id, nn.id) : (nn.id, a.id)
     if !(new_pair in pairs)
       # We also need to check if our current pair is closer to each
       # other than any pair using our first id already in the list,
       # so we keep track of nn distances.
-      dist = pair_distance(a.pos, model[nid].pos, model.space.metric)
+      dist = pair_distance(a.pos, model[nn.id].pos, model.space.metric)
 
       idx = findfirst(x->first(new_pair) == x, first.(pairs))
       if idx == nothing
@@ -406,7 +375,7 @@ function true_pairs!(pairs::Vector{Tuple{Int, Int}}, model::ABM, r)
   end
 end
 
-function pair_distance(pos1, pos2, metric)
+function pair_distance(pos1, pos2, metric::Symbol)
   if metric == :euclidean
     sqrt(sum(abs2.(pos1 .- pos2)))
   elseif metric == :cityblock