--- a/test/continuousSpace_tests.jl
+++ b/test/continuousSpace_tests.jl
@@ -105,4 +105,7 @@ end
   for id in fi
     @test id ∉ se
   end
+  pairs = interacting_pairs(model, 2.0; method = :all).pairs
+  @test length(pairs) == 5
+  @test (3, 6) ∉ pairs
 end