WIP for discussions around #220.

This implements the *all neighbors* scenario at the moment, with a currently unimplemented *true neighbors* to come soon. 

Current benchmarks using test case:
```julia
julia> @btime interacting_pairs(model, 2.0).pairs
  5.489 μs (73 allocations: 3.41 KiB)
2-element Array{Tuple{Int64,Int64},1}:
 (6, 5)
 (4, 3)

julia> @btime interacting_pairs(model, 2.0; method = :all).pairs
  10.959 μs (137 allocations: 6.47 KiB)
5-element Array{Tuple{Int64,Int64},1}:
 (3, 4)
 (4, 5)
 (4, 6)
 (3, 5)
 (5, 6)
```