@Datseris 2020-04-17
These benchmarks are not really realistic and some where 100s or 1000s of agents exist in the database should be made. I Think this simple factor of 2 difference ins not true.