@Libbum 2020-04-17
Current implementation:

```julia
using Agents, Random
Random.seed!(6548)
space = ContinuousSpace(2, extend = (10, 10), periodic = false, metric = :euclidean)
model = ABM(Agent6, space; scheduler = by_id)
for i in 1:10000                                                                                                                                                            
    add_agent_pos!(Agent6(i, (50*rand(), 50*rand()), (0.0, 0.0), 0), model)                                                                                                               
end 

julia> @btime interacting_pairs(model, 2.0).pairs
  5.346 s (2088735 allocations: 54.53 MiB)
  650-element Array{Tuple{Int64,Int64},1}:
  
julia> @btime interacting_pairs(model, 2.0; method = :true).pairs
  5.491 s (2154985 allocations: 95.69 MiB)
  650-element Array{Tuple{Int64,Int64},1}:
  
julia> @btime interacting_pairs(model, 2.0; method = :all).pairs
  62.462 s (2144994 allocations: 64.81 MiB)
307203-element Array{Tuple{Int64,Int64},1}:
```