@Datseris 2020-04-17
Seems good to me. Let's wait for feedback from @lhupe before merging this.