@Datseris 2020-04-17
Oh wow:
```
julia> @btime interacting_pairs(model, 2.0).pairs
  5.346 s (2088735 allocations: 54.53 MiB)
  650-element Array{Tuple{Int64,Int64},1}:
  
julia> @btime interacting_pairs(model, 2.0; method = :true).pairs
  5.491 s (2154985 allocations: 95.69 MiB)
  650-element Array{Tuple{Int64,Int64},1}:
```
if the `true` method is as fast as the scheduled one maybe it should be the default.