@Libbum 2020-04-17
The question I have would be: Is there a reason to actually keep `:scheduler`? I don't really see a use case for it, but perhaps I overlook something.

If there's no reason to keep it, we can change the default to the current `:true` implementation, then remove the `method` keyword replacing it with an `all = false` keyword.