@Datseris 2020-04-19
@Libbum alright, let's go with the change you suggested and make `all = false` the default keyword.