@Libbum 2020-04-19
028d2e3 drops the scheduler method as discussed, and fixes a bug when there are no nearest neighbors.

There are currently three broken collision tests to be checked before merging. They pass if we now use `interacting_pairs(; all = true)`, but I'm unsure if this is the behaviour the tests expect, or they if fact rely on the `:scheduler` implementation.