@Libbum 2020-04-19
The collision tests are for a small radius and essentially test for the dynamics in the social distancing example. Setting all = true makes everything work again. I can just push that fix directly to master.