@Datseris 2020-04-19
It is definitely incorrect to use `all` in social distancing. The dynamics are billiard ball collisions. It makes no sense for an agent to interact with more than one in one step.